+++ src/handler.ts
@@ -13,6 +13,13 @@ const parseSetCookies = (headers: Headers, setCookie: string | string[]) => {
 		for (let i = 0; i < setCookie.length; i++) {
 			const index = setCookie[i].indexOf('=')
 
+			console.log(
+				'Append',
+				`${setCookie[i].slice(0, index)}=${setCookie[i].slice(
+					index + 1
+				)}`
+			)
+
 			headers.append(
 				'Set-Cookie',
 				`${setCookie[i].slice(0, index)}=${setCookie[i].slice(
@@ -173,7 +180,13 @@ export const mapResponse = (
 
 					default:
 						if (!set.headers['Content-Type'])
-							set.headers['Content-Type'] = 'application/json'
+							if (set.headers instanceof Headers)
+								set.headers.append(
+									'Content-Type',
+									'application/json'
+								)
+							else
+								set.headers['Content-Type'] = 'application/json'
 
 						return new Response(JSON.stringify(response), {
 							status: set.status,